ang parameters are integers not reals

+++ src/input/openrsp_cfg.F90
@@ -81,7 +81,7 @@ module openrsp_cfg
    ! keywords that control the XCint grid (not the Dalton grid)
    logical :: openrsp_cfg_use_xcint_grid = .false.
    real(8) :: openrsp_cfg_radint = 1.0d-12
-   real(8) :: openrsp_cfg_angmin = 86
-   real(8) :: openrsp_cfg_angint = 302
+   integer :: openrsp_cfg_angmin = 86
+   integer :: openrsp_cfg_angint = 302
 
 end module
